refactor(banners): move add, update, deactivate logic to admin controller/services
